+++ src/serenity/mdrecorder/phemex.py
@@ -9,7 +9,7 @@ from tau.core import MutableSignal, NetworkScheduler
 from tau.event import Do
 from tau.signal import Map, Filter
 
-from serenity.mdrecorder.journal import Journal
+from serenity.tickstore.journal import Journal
 from serenity.mdrecorder.subscriber import subscribe_trades
 from serenity.utils import init_logging
 from pathlib import Path