+++ HISTORY.rst
@@ -3,6 +3,13 @@
 Release History
 ---------------
 
+0.0.3 (2020-04-19)
++++++++++++++++++++
+
+- Added new generic FeedHandler API
+- Implemented PhemexFeedHandler
+- Moved Tickstore and Journal from serenity.mdrecorder to serenity.tickstore
+
 0.0.2 (2020-04-13)
 +++++++++++++++++++